test: isolate resend cache integration coverage
