+++ crates/atm-daemon/src/tests/runtime_root.rs
@@ -12,10 +12,7 @@ use atm_core::types::CommandAction;
 use atm_core::types::ReadSelection;
 use atm_core::{ApiRequest, ApiRouter, AuthenticatedIngress, RequestDeadline};
 use atm_runtime_test_support::{SQLITE_RUNTIME_PATH_ENV, open_sqlite_boundary};
-use atm_storage::{
-    HostName, HttpsInterface, MessageKey, MessageQuery, PeerAliasKey, PeerResendCacheSetting,
-    TrustedPeer,
-};
+use atm_storage::{HostName, HttpsInterface, MessageKey, MessageQuery, PeerAliasKey, TrustedPeer};
 use std::io::Write;
 use std::net::{IpAddr, Ipv4Addr, Ipv6Addr, TcpListener};
 use std::num::NonZeroU16;
@@ -30,7 +27,7 @@ use crate::test_support::{
     write_test_local_ipc_request,
 };
 
-fn add_member_via_retained_admin(
+pub(super) fn add_member_via_retained_admin(
     db_path: &std::path::Path,
     atm_home: &std::path::Path,
     team: &str,
@@ -55,7 +52,11 @@ fn add_member_via_retained_admin(
     .expect("add member");
 }
 
-fn configure_trusted_peer(db_path: &std::path::Path, canonical_host: &str, aliases: &[&str]) {
+pub(super) fn configure_trusted_peer(
+    db_path: &std::path::Path,
+    canonical_host: &str,
+    aliases: &[&str],
+) {
     assert!(
         canonical_host.parse::<IpAddr>().is_err(),
         "test canonical peer hosts must be stable DNS names; use an IP alias instead"
@@ -81,7 +82,7 @@ fn configure_trusted_peer(db_path: &std::path::Path, canonical_host: &str, alias
     }
 }
 
-fn configure_peer_http_source(db_path: &std::path::Path) {
+pub(super) fn configure_peer_http_source(db_path: &std::path::Path) {
     let assembly = open_sqlite_boundary(db_path).expect("sqlite boundary");
     assembly
         .peer_config_store()
@@ -198,23 +199,7 @@ fn serve_failed_then_pipelined_peer_responses(
     (port, server)
 }
 
-fn serve_dropped_peer_requests(request_count: usize) -> (NonZeroU16, thread::JoinHandle<()>) {
-    let listener = TcpListener::bind((Ipv6Addr::LOCALHOST, 0)).expect("peer listener");
-    let port = NonZeroU16::new(listener.local_addr().expect("peer address").port())
-        .expect("non-zero peer port");
-    let server = thread::spawn(move || {
-        for _ in 0..request_count {
-            let (mut stream, _) = listener.accept().expect("peer accept");
-            let _ = HttpFrameReader::new()
-                .read_request(&mut stream)
-                .expect("read peer request")
-                .expect("one peer request");
-        }
-    });
-    (port, server)
-}
-
-fn remote_write_request(
+pub(super) fn remote_write_request(
     atm_home: std::path::PathBuf,
     workspace_dir: std::path::PathBuf,
     body: String,
@@ -402,89 +387,6 @@ fn failed_direct_peer_response_retains_the_unconfirmed_marker() {
             .and_then(|value| value.get("host"))
             .and_then(serde_json::Value::as_str),
         Some("localhost")
-    );
-    server.join().expect("peer server");
-}
-
-#[test]
-#[serial_test::serial(env)]
-fn resend_cache_reload_keeps_disabled_delivery_on_the_direct_no_retry_path() {
-    install_retained_runtime_factory();
-    let tempdir = TempDir::new().expect("tempdir");
-    let atm_home = tempdir.path().join("atm-home");
-    let workspace_dir = tempdir.path().join("workspace");
-    std::fs::create_dir_all(&atm_home).expect("atm home dir");
-    std::fs::create_dir_all(&workspace_dir).expect("workspace dir");
-    let db_path = tempdir.path().join("mail.db");
-    write_team_config(&atm_home, &[]);
-    add_member_via_retained_admin(
-        &db_path,
-        &atm_home,
-        TEST_TEAM,
-        ROLE_TEAM_LEAD,
-        &workspace_dir,
-    );
-    let (port, server) = serve_dropped_peer_requests(1);
-    configure_trusted_peer(&db_path, "localhost", &["peer.example.test", "127.0.0.1"]);
-    let assembly = open_sqlite_boundary(&db_path).expect("sqlite boundary");
-    assembly
-        .peer_config_store()
-        .save_trusted_peer(&TrustedPeer {
-            host: "localhost".parse().expect("canonical host"),
-            fingerprint: "sha256:test".parse().expect("fingerprint"),
-            enabled: true,
-            https_port: port,
-        })
-        .expect("replace peer endpoint with test listener");
-    assembly
-        .peer_config_store()
-        .save_peer_resend_cache_setting(PeerResendCacheSetting { enabled: false })
-        .expect("disable cache");
-    configure_peer_http_source(&db_path);
-    let dispatcher = DaemonRequestDispatcher::new_for_test(
-        atm_home.clone(),
-        RuntimeStatusCache::new(),
-        db_path.clone(),
-    );
-
-    let disabled_error = dispatcher
-        .dispatch(remote_write_request(
-            atm_home.clone(),
-            workspace_dir.clone(),
-            "disabled direct peer write".to_owned(),
-        ))
-        .expect_err("disabled caching returns the direct peer failure");
-    assert_eq!(
-        disabled_error.code(),
-        AtmErrorCode::RemoteDeliveryUnconfirmed
-    );
-    assert_eq!(
-        dispatcher.next_peer_resend_due(),
-        None,
-        "disabled caching creates neither an aggregate nor a due event"
-    );
-
-    assembly
-        .peer_config_store()
-        .save_peer_resend_cache_setting(PeerResendCacheSetting { enabled: true })
-        .expect("enable cache");
-    dispatcher
-        .reload_runtime_view()
-        .expect("reload enabled cache");
-    let enabled_error = dispatcher
-        .dispatch(remote_write_request(
-            atm_home,
-            workspace_dir,
-            "enabled direct peer write".to_owned(),
-        ))
-        .expect_err("enabled cache leaves the new admission queued behind retained work");
-    assert_eq!(
-        enabled_error.code(),
-        AtmErrorCode::RemoteDeliveryUnconfirmed
-    );
-    assert!(
-        dispatcher.next_peer_resend_due().is_some(),
-        "enabled reload bootstraps retained work into one later recovery attempt"
     );
     server.join().expect("peer server");
 }
